Support processors, introduce :content key, switch to keyword keys
This commit introduces a number of features and breaks compability with
v0.1.1.

First, it introduces a processing step to `Feedstock.format_content`
that allows a user to use a lambda (called a processor in Feedstock) to
process the text extracted from the page. If no processor is provided,
the text is passed through as-is.

Second, rather than using a combination of keys to determine how to
extract the content from a node, Feedstock now supports the `:content`
key. The `:content` key can be a `{ attribute: <attribute_name> }` hash,
the `"inner_content"` or left unset. Feedstock uses Ruby's new pattern
matching feature to support this and so the minimum Ruby version
required is now 2.7.

Third, Feedstock now requires the keys provided in a rules hash to be
keywords, not strings. This is a requirement for pattern matching.