Support parsing XML and filtering entries
This commit adds support for two things: (1) creating a feed from an XML
file and (2) filtering entries according to a provided filter function.

Together, these changes allow Feedstock to be used to generate feeds
_from_ other feeds. An example usage might be where a user wishes to
filter a podcast feed to only include certain episodes.

To support ingesting XML files, this commit adds an additional argument
to the `Feedstock.feed` method. This is a breaking change from v0.2.